Edited frontend to take examples to a different page
